supprimer commencement

+++ sachem/Controllers/EtudiantController.cs
@@ -192,26 +192,38 @@ namespace sachem.Controllers
         public ActionResult DeleteConfirmed(int id,int? page)
         {
             var pageNumber = page ?? 1;
-            
+       
             Personne personne = db.Personne.Find(id);
+            var inscription = db.Inscription.Where(x => x.id_Pers == personne.id_Pers).FirstOrDefault();
 
-            var etuProgEtu = db.EtuProgEtude.Where(x => x.id_Etu == personne.id_Pers);
-            db.EtuProgEtude.RemoveRange(etuProgEtu);
-            var groupeEtu = db.GroupeEtudiant.Where(y => y.id_Etudiant == personne.id_Pers);
-            db.GroupeEtudiant.RemoveRange(groupeEtu);
-            var Jumul = db.Jumelage.Where(z => z.id_InscEleve == personne.id_Pers);
-            db.Jumelage.RemoveRange(Jumul);
-            var Inscri = db.Inscription.Where(a => a.id_Pers == personne.id_Pers);
-            db.Inscription.RemoveRange(Inscri);
-            var CoursSuiv = db.CoursSuivi.Where(b => b.id_Pers == personne.id_Pers);
-            db.CoursSuivi.RemoveRange(CoursSuiv);
+            if (db.GroupeEtudiant.Any(x => x.id_Etudiant == personne.id_Pers))
+                ModelState.AddModelError(string.Empty, Messages.I_014());
+            if (inscription != null)
+            {
+                if (db.Jumelage.Any(x => x.id_InscEleve == inscription.id_Inscription))
+                    ModelState.AddModelError(string.Empty, Messages.I_043());//changer erreur
+            }
+            if (ModelState.IsValid)
+            {
+                var etuProgEtu = db.EtuProgEtude.Where(x => x.id_Etu == personne.id_Pers);
+                db.EtuProgEtude.RemoveRange(etuProgEtu);
+                var groupeEtu = db.GroupeEtudiant.Where(y => y.id_Etudiant == personne.id_Pers);
+                db.GroupeEtudiant.RemoveRange(groupeEtu);
+                var Jumul = db.Jumelage.Where(z => z.id_InscEleve == personne.id_Pers);
+                db.Jumelage.RemoveRange(Jumul);
+                var Inscri = db.Inscription.Where(a => a.id_Pers == personne.id_Pers);
+                db.Inscription.RemoveRange(Inscri);
+                var CoursSuiv = db.CoursSuivi.Where(b => b.id_Pers == personne.id_Pers);
+                db.CoursSuivi.RemoveRange(CoursSuiv);
+                db.Personne.Remove(personne);
+                db.SaveChanges();
+                TempData["Success"] = Messages.I_028(personne.NomPrenom);
+                //redirection à l'index après la suppression
+            }
+            else { }
 
-            //suppresion et sauvegarde dans la bd
-            db.Personne.Remove(personne);
-            db.SaveChanges();
-            TempData["Success"] = Messages.I_028(personne.NomPrenom);
-            //redirection à l'index après la suppression
             return RedirectToAction("Index");
+
         }
         //fonction qui supprime un programme d'étude à oartir de la page modifier
         [ValidationAccesEnseignant]
@@ -232,6 +244,7 @@ namespace sachem.Controllers
             // Verifier si le matricule existe déja dans la BD
             if (db.Personne.Any(x => x.Matricule == personne.Matricule))
                 ModelState.AddModelError(string.Empty, Messages.I_004(personne.Matricule));
+
         }
         protected override void Dispose(bool disposing)
         {
